+++ src/components/media/Image.vue
@@ -6,12 +6,13 @@
         <img :src="image.imageURL" width="100%" />
       </div>
       <div class="col-12" v-if="image.tags.length > 0 || image.bibleRefs.length > 0">
-        <q-chip v-if="image.tags.length > 0" v-for="tag in image.tags" :key="tag">{{ tag }}</q-chip>
-        <!-- |
-        <q-chip v-if="image.bibleRefs.length > 0" v-for="bibleRef in image.bibleRefs" :key="bibleRef">{{ bibleRef }}</q-chip> -->
+        <q-chip v-if="image.tags.length > 0" v-for="tag in image.tags" :key="tag" color="primary">{{ tag }}</q-chip>
+        |
+        <q-chip v-if="readableRefs.length > 0" v-for="ref in readableRefs" :key="ref" color="secondary">{{ ref }}</q-chip>
       </div>
       <div class="col-12">
         <q-btn color="primary" @click.native="editing = true">Edit</q-btn>
+        <q-btn color="primary" v-if="addModule !== undefined" @click.native="add">Add</q-btn>
       </div>
     </div>
     <div class="row gutter-sm" v-if="editing">
@@ -25,6 +26,14 @@
         <q-chips-input v-model="image.tags" float-label="Tags" />
       </div>
       <div class="col-12">
+        <q-chips-input
+          color="secondary"
+          v-model="readableRefs"
+          float-label="Bible References"
+          @input="addRef"
+        />
+      </div>
+      <div class="col-12">
         <q-btn color="primary" @click.native="save">Save</q-btn>
       </div>
     </div>
@@ -35,7 +44,7 @@
 import { Notify } from 'quasar'
 
 export default {
-  props: ['data', 'open', 'close'],
+  props: ['data', 'open', 'close', 'addModule'],
   name: 'media-image',
   data () {
     return {
@@ -43,7 +52,8 @@ export default {
       image: {
         tags: [],
         bibleRefs: []
-      }
+      },
+      readableRefs: []
     }
   },
   watch: {
@@ -52,13 +62,21 @@ export default {
       if (val !== {}) {
         this.editing = false
         this.image = val
+        this.readableRefs = val.bibleRefs.map(e => { return this.$bible.readable(e) })
       }
     }
   },
   mounted () {
-    this.image = this.data
+    this.init()
   },
   methods: {
+    init () {
+      this.image = this.data
+    },
+    addRef (newRef) {
+      this.image.bibleRefs = newRef.map(e => { return this.$bible.parse(e) })
+      this.readableRefs = newRef.map(e => { return this.$bible.readable(e) })
+    },
     save () {
       this.$database.update('image', this.image._id, this.image, (res) => {
         Notify.create({
@@ -68,6 +86,10 @@ export default {
         })
         this.editing = false
       })
+    },
+    add () {
+      this.addModule(this.image._id, 'image', this.image)
+      this.close()
     }
   }
 }